Make include paths absolute
Paths in included files need to be absolute (because relative paths
would be resolved from wherever the files are included from), and
since other files might get moved around it's easier to have all paths
be absolute. The only sacrifice is portability of the root folder, but
I don't imagine that will ever be a problem with this website.